State value is float64

+++ api.go
@@ -324,7 +324,7 @@ func getPartyShaders(c *gin.Context) {
                 for hash, value := range hashData {
                     // We will track the counts of shader hashes present using a hash
                     // If the hash value int is equal to the number of characters in the party, then everyone has the shader
-                    state := value.(map[string]interface{})["state"].(int)
+                    state := value.(map[string]interface{})["state"].(float64)
                     if state != 0 {
                         continue
                     }
